# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Postaway",
    "description": "Postaway - Social media app",
    "version": "1.0.0"
  },
  "host": "localhost:3500",
  "securityDefinitions": {
    "JWT": {
      "in": "header",
      "name": "authorization",
      "type": "apiKey"
    }
  },
  "paths": {
...
    "/api/posts": {
      "get": {
        "tags": ["Posts"],
        "summary": "Get Posts by User Id",
        "description": "Returns all the posts, uploaded by the current logged in user",
        "produces": "application/json",
        "security": [
          {
            "JWT": {}
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "An array containing post objects",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Post"
              }
            }
          },
          "404": {
            "description": "Not Found Error",
            "schema": {
              "$ref": "#/definitions/NotFoundError"
            }
          },
          "401": {
            "description": "Authorization Error",
            "schema": {
              "$ref": "#/definitions/AuthorizationError"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/InternalServerError"
            }
          }
        }
      },
      "post": {
        "tags": ["Posts"],
        "summary": "Upload a post",
        "description": "Uploads a post(caption and image) onto the server",
        "security": [{ "JWT": {} }],
        "produces": "application/json",
        "consumes": ["multipart/form-data"],
        "parameters": [
          {
            "in": "formData",
            "name": "caption",
            "requied": true,
            "type": "string",
            "description": "Caption of the post that needs to be uploaded",
            "example": "My new Post"
          },
          {
            "in": "formData",
            "name": "imageURL",
            "type": "file",
            "required": true,
            "description": "The file to upload"
          }
        ],
        "responses": {
          "201": {
            "description": "Created Post Object",
            "schema": {
              "$ref": "#/definitions/Post"
            }
          },
          "302": {
            "description": "Validation Error",
            "schema": {
              "$ref": "#/definitions/ValidationError"
            }
          },
          "401": {
            "description": "Authorization Error",
            "schema": {
              "$ref": "#/definitions/AuthorizationError"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
    "/api/posts/{id}": {
      "get": {
        "tags": ["Posts"],
        "summary": "Get Posts by Post Id",
        "description": "Retrieve a particular post by it's id",
        "security": [
          {
            "JWT": {}
          }
        ],
        "produces": "application/json",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "description": "ID of the post to retrieve",
            "example": 2
          }
        ],
        "responses": {
          "200": {
            "description": "Post Object",
            "schema": {
              "$ref": "#/definitions/Post"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "404": {
            "description": "Not Found Error",
            "schema": {
              "$ref": "#/definitions/NotFoundError"
            }
          },
          "401": {
            "description": "Authorization Error",
            "schema": {
              "$ref": "#/definitions/AuthorizationError"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/InternalServerError"
            }
          }
        }
      },
      "delete": {
        "tags": ["Posts"],
        "summary": "Delete a post",
        "description": "Delete a post by it's id",
        "security": [{ "JWT": {} }],
        "produces": "application/json",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Id of the post that needs to be deleted",
            "type": "integer",
            "example": 2
          }
        ],
        "responses": {
          "200": {
            "description": "Object containing success message",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": "true"
                },
                "message": {
                  "type": "string",
                  "example": "Post with post id {id} deleted successfully"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "404": {
            "description": "Not Found Error",
            "schema": {
              "$ref": "#/definitions/NotFoundError"
            }
          },
          "401": {
            "description": "Authorization Error",
            "schema": {
              "$ref": "#/definitions/AuthorizationError"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/InternalServerError"
            }
          }
        }
      },
      "put": {
        "tags": ["Posts"],
        "summary": "Update a post by id",
        "description": "Update a post by postId, provided in the path",
        "security": [{ "JWT": {} }],
        "produces": "application/json",
        "consumes": ["multipart/form-data"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "requied": true,
            "type": "integer",
            "description": "Id of the post that needs to be updated",
            "exapmle": "2"
          },
          {
            "in": "formData",
            "name": "caption",
            "requied": false,
            "type": "string",
            "description": "Caption of the post that needs to be uploaded",
            "exapmle": "My new Post"
          },
          {
            "in": "formData",
            "name": "imageURL",
            "type": "file",
            "required": false,
            "description": "The file to upload"
          }
        ],
        "responses": {
          "201": {
            "description": "Updated post object",
            "schema": {
              "$ref": "#/definitions/Post"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/BadRequestError"
            }
          },
          "404": {
            "description": "Not Found Error",
            "schema": {
              "$ref": "#/definitions/NotFoundError"
            }
          },
          "401": {
            "description": "Authorization Error",
            "schema": {
              "$ref": "#/definitions/AuthorizationError"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "ValidationError": {
      "type": "object",
      "properties": {
        "error": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "type": { "type": "string", "example": "field" },
              "value": { "type": "string", "example": "value" },
              "msg": {
                "type": "string",
                "example": "Password must be at least 8 characters long"
              },
              "path": { "type": "string", "example": "password" },
              "location": { "type": "string", "example": "body" }
            }
          }
        },
        "name": {
          "type": "string",
          "example": "ValidationError"
        },
        "stack": {
          "type": "string"
        }
      }
    },
    "NotFoundError": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "stack": {
          "type": "string"
        }
      }
    },
    "AuthorizationError": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "stack": {
          "type": "string"
        }
      }
    },
    "InternalServerError": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "stack": {
          "type": "string"
        }
      }
    },
    "BadRequestError": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "example": "Id is invalid"
        },
        "name": {
          "type": "string",
          "example": "Validation Error"
        },
        "stack": {
          "type": "string"
        }
      }
    },
...
    "Post": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "userId": {
          "type": "integer"
        },
        "caption": {
          "type": "string"
        },
        "imageURL": {
          "type": "string"
        },
        "comments": {
          "type": "integer"
        },
        "likes": {
          "type": "integer"
        },
        "status": {
          "type": "string"
        },
        "createdAt": {
          "type": "integer"
        }
      }
    },
...
  }
}